Move random_tensornetwork, random_ttn, random_mps, ModelHamiltonians to test/utils.jl
Pull non-basic constructors that are only used by tests out of `src/` and
into a new `test/utils.jl` helper file. Each test that needs them does
`include("utils.jl")` inside its gensym module. The file is intentionally
named `utils.jl` so the `ITensorPkgSkeleton.runtests` runner does not
auto-pick it up (the runner only globs `test_*.jl`).

What moved:
- `random_tensornetwork` (12 method specializations, including the
  `Distribution`- and `IsUnderlyingGraph`-based variants) — moved from
  `src/specialitensornetworks.jl` to `test/utils.jl`.
- `random_ttn` and `random_mps` — moved from
  `src/treetensornetworks/treetensornetwork.jl` to `test/utils.jl`.
  These thin wrappers around `random_tensornetwork` had no real consumers
  in `src/` (only docstring examples) and a single test reference.
- `ModelHamiltonians.tight_binding`, `heisenberg`, `ising`, plus the
  internal helpers `to_callable`, `nth_nearest_neighbors`,
  `next_nearest_neighbors` — moved from `src/lib/ModelHamiltonians/` to
  a `ModelHamiltonians` submodule in `test/utils.jl`. Test sites continue
  to call them as `ModelHamiltonians.ising(...)` etc.

What was deleted outright:
- `delta_network` (and the entire `src/specialitensornetworks.jl`) — the
  only consumer was `ModelNetworks.ising_network`, which itself had no
  consumers.
- `ModelHamiltonians.hubbard` — no consumers.
- `src/lib/ModelNetworks/` (`ising_network`, `ising_network_state`) — no
  consumers; the lone `using ITensorNetworks.ModelNetworks: ModelNetworks`
  in `test/test_belief_propagation.jl` was an unused import.
- `random_ttn` / `random_mps` jldoctest examples in `src/normalize.jl`,
  `src/expect.jl`, `src/treetensornetworks/abstracttreetensornetwork.jl`
  (per the rule that doc examples should not depend on functions being
  removed). The `[`random_ttn`](@ref)` / `[`random_mps`](@ref)` references
  in `See also:` lists were dropped as well.

`docs/src/computing_properties.md`, `tree_tensor_networks.md`, and
`solvers.md` were updated to drop their `random_ttn` / `random_mps` calls;
the `@autodocs` listing in `docs/src/reference.md` was trimmed to just
`ITensorNetworks` (no longer pulling in the removed `ModelNetworks` /
`ModelHamiltonians` submodules).

A few imports that were previously available inside `ITensorNetworks` as
a side effect of the deleted files' `using` clauses are now imported
directly by the files that actually use them:
- `ITensors: delta` — `src/formnetworks/bilinearformnetwork.jl`,
  `src/caches/abstractbeliefpropagationcache.jl`

Bump to 0.19.0 (breaking, pre-1.0 minor bump) and update pinned
`ITensorNetworks` compat in `test/`, `docs/`, and `examples/` Project.toml.
Also add `SimpleTraits` to `test/Project.toml` so `test/utils.jl` can use
the `@traitfn` macro.
