+++ docs/Project.toml
@@ -16,7 +16,7 @@ path = ".."
 Documenter = "1.10"
 Graphs = "1"
 ITensorFormatter = "0.2.27"
-ITensorNetworks = "0.18"
+ITensorNetworks = "0.19"
 ITensors = "0.9"
 Literate = "2.20.1"
 NamedGraphs = "0.8.2"
