+++ src/ITensorNetworks.jl
@@ -14,7 +14,6 @@ include("contraction_sequences.jl")
 include("tebd.jl")
 include("itensornetwork.jl")
 include("contract.jl")
-include("specialitensornetworks.jl")
 include("partitioneditensornetwork.jl")
 include("edge_sequences.jl")
 include("caches/abstractbeliefpropagationcache.jl")
@@ -63,7 +62,5 @@ include("normalize.jl")
 include("expect.jl")
 include("environment.jl")
 include("exports.jl")
-include("lib/ModelHamiltonians/src/ModelHamiltonians.jl")
-include("lib/ModelNetworks/src/ModelNetworks.jl")
 
 end
