+++ src/formnetworks/bilinearformnetwork.jl
@@ -1,6 +1,6 @@
 using Adapt: adapt
 using ITensors.NDTensors: datatype, denseblocks
-using ITensors: ITensor, Op, prime, sim
+using ITensors: ITensor, Op, delta, prime, sim
 using NamedGraphs.GraphsExtensions: disjoint_union
 
 default_dual_site_index_map = prime
