+++ src/normalize.jl
@@ -48,25 +48,6 @@ multiplied by the same scalar factor).
   - `alg="exact"`: Normalization algorithm. `"exact"` contracts ⟨ψ|ψ⟩ exactly;
     `"bp"` uses belief propagation for large networks.
 
-# Example
-
-```jldoctest
-julia> using NamedGraphs.NamedGraphGenerators: named_grid
-
-julia> using LinearAlgebra: norm
-
-julia> g = named_grid((4,));
-
-julia> s = siteinds("S=1/2", g);
-
-julia> psi = random_ttn(s; link_space = 2);
-
-julia> psi = normalize(psi);
-
-julia> norm(psi) ≈ 1
-true
-```
-
 See also: `norm`, [`inner`](@ref ITensorNetworks.inner).
 """
 function LinearAlgebra.normalize(tn::AbstractITensorNetwork; alg = "exact", kwargs...)
