+++ test/Project.toml
@@ -24,6 +24,7 @@ OrdinaryDiffEqTsit5 = "b1df2697-797e-41e3-8120-5422d3b24e4a"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
+SimpleTraits = "699a6c99-e7fa-54fc-8d76-47d257e15c1d"
 SplitApplyCombine = "03a91e81-4c3e-53e1-a0a4-9c0c8f19dd66"
 StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
@@ -46,7 +47,7 @@ Glob = "1.3.1"
 Graphs = "1.12"
 GraphsFlows = "0.1.1"
 ITensorMPS = "0.3.6"
-ITensorNetworks = "0.18"
+ITensorNetworks = "0.19"
 ITensorPkgSkeleton = "0.3.42"
 ITensors = "0.7, 0.8, 0.9"
 KrylovKit = "0.8, 0.9, 0.10"
@@ -60,6 +61,7 @@ OrdinaryDiffEqTsit5 = "1.1"
 Pkg = "1.10"
 Random = "1.10"
 SafeTestsets = "0.1"
+SimpleTraits = "0.9.4"
 SplitApplyCombine = "1.2.3"
 StableRNGs = "1.0.2"
 Suppressor = "0.2.8"
