+++ test/test_expect.jl
@@ -1,13 +1,14 @@
 @eval module $(gensym())
 using Graphs: SimpleGraph, uniform_tree
-using ITensorNetworks: BeliefPropagationCache, ITensorNetwork, expect,
-    original_state_vertex, random_tensornetwork, siteinds
+using ITensorNetworks:
+    BeliefPropagationCache, ITensorNetwork, expect, original_state_vertex, siteinds
 using NamedGraphs.NamedGraphGenerators: named_grid
 using NamedGraphs: NamedGraph, vertices
 using SplitApplyCombine: group
 using StableRNGs: StableRNG
 using TensorOperations: TensorOperations
 using Test: @test, @testset
+include("utils.jl")
 @testset "Test Expect" begin
     #Test on a tree
     L, χ = 4, 2
