+++ test/test_itensornetwork.jl
@@ -6,8 +6,7 @@ using Graphs: degree, dijkstra_shortest_paths, edges, grid, has_vertex, ne, neig
 using GraphsFlows: GraphsFlows
 using ITensorNetworks: ITensorNetworks, ITensorNetwork, IndsNetwork, contraction_sequence,
     flatten_linkinds, flatten_siteinds, inner_network, linkinds, neighbor_tensors, norm_sqr,
-    norm_sqr_network, orthogonalize, random_tensornetwork, siteinds, tree_orthogonalize,
-    ttn, ⊗
+    norm_sqr_network, orthogonalize, siteinds, tree_orthogonalize, ttn, ⊗
 using ITensors.NDTensors: NDTensors, dim
 using ITensors: ITensors, ITensor, Index, Op, commonind, commoninds, contract, dag,
     hascommoninds, hasinds, inds, inner, itensor, onehot, order, prime, random_itensor,
@@ -21,6 +20,7 @@ using Random: randn!
 using StableRNGs: StableRNG
 using TensorOperations: TensorOperations
 using Test: @test, @test_broken, @testset
+include("utils.jl")
 const elts = (Float32, Float64, Complex{Float32}, Complex{Float64})
 @testset "ITensorNetwork tests" begin
     @testset "ITensorNetwork Basics" begin
