+++ test/test_tebd.jl
@@ -1,12 +1,13 @@
 @eval module $(gensym())
 using Graphs: vertices
 using ITensorNetworks.ITensorsExtensions: group_terms
-using ITensorNetworks.ModelHamiltonians: ModelHamiltonians
 using ITensorNetworks: ITensorNetwork, cartesian_to_linear, dmrg, expect, siteinds, tebd
 using ITensors: ITensors
 using NamedGraphs.GraphsExtensions: rename_vertices
 using NamedGraphs.NamedGraphGenerators: named_grid
 using Test: @test, @test_broken, @testset
+include("utils.jl")
+using .ModelHamiltonians: ModelHamiltonians
 
 ITensors.disable_warn_order()
