+++ test/test_ttn_expect.jl
@@ -1,6 +1,6 @@
 @eval module $(gensym())
 using Graphs: vertices
-using ITensorNetworks: expect, random_mps, siteinds, ttn
+using ITensorNetworks: expect, siteinds, ttn
 using LinearAlgebra: norm
 using NamedGraphs.NamedGraphGenerators: named_comb_tree
 using StableRNGs: StableRNG
